+++ main.py
@@ -12,6 +12,9 @@ from vertexai.generative_models import (
     Part,
 )
 
+from google import genai
+from google.genai.types import EmbedContentConfig
+
 from helpers.image_utils import image_resize
 from helpers.recommender_utils import any_list_element_in_string
 from helpers.recommender_utils import get_cosine_score
@@ -109,8 +112,18 @@ def get_text_embedding_from_text_embedding_model(
                                The format (list or NumPy array) depends on the
                                value of the 'return_array' parameter.
     """
-    embeddings = text_embedding_model.get_embeddings([text])
-    text_embedding = [embedding.values for embedding in embeddings][0]
+    client = genai.Client()
+    response = client.models.embed_content(
+        #model="text-embedding-005",
+        model="text-embedding-large-exp-03-07",
+        contents=text,
+        config=EmbedContentConfig(
+            task_type="RETRIEVAL_DOCUMENT",  # Optional
+            output_dimensionality=768,  # Optional
+        ),
+    )
+
+    text_embedding = response.embeddings[0].values
 
     if return_array:
         text_embedding = np.fromiter(text_embedding, dtype=float)
@@ -228,8 +241,9 @@ def get_reference_image_description(image_filename: str) -> list:
 
   response = multimodal_model.generate_content(
     [
-      "Can you describe the clothes in the photo, including style, color, and any designs?  Make sure to only describe each individual article of clothing, and give a separate response.",
-       image
+        #"Can you describe the clothes in the photo, including style, color, and any designs?  Make sure to only describe each individual article of clothing, and give a separate response.",
+        "Describe the clothes in the photo.  Description should only include clothing 'type', 'color', 'style', and 'design'. Make sure to only describe each individual article of clothing, and give a separate response.",
+        image
     ],
     generation_config=generation_config
   )
@@ -251,9 +265,7 @@ from vertexai.vision_models import MultiModalEmbeddingModel
 # for embedding
 text_embedding_model = TextEmbeddingModel.from_pretrained("text-embedding-005")
 
-#image_metadata_df_csv = pd.read_csv("mywardrobe_1-0-pro-vision.csv",converters={"image_description_text_embedding": lambda x: x.strip("[]").split(", ")})
-#image_metadata_df_csv = pd.read_csv("mywardrobe_1-5-pro.csv",converters={"image_description_text_embedding": lambda x: x.strip("[]").split(", ")})
-image_metadata_df_csv = pd.read_csv("mywardrobe_2-0-flash.csv",converters={"image_description_text_embedding": lambda x: x.strip("[]").split(", ")})
+image_metadata_df_csv = pd.read_csv("mywardrobe.csv",converters={"image_description_text_embedding": lambda x: x.strip("[]").split(", ")})
 
 
 #--------------